+++ CHANGES.md
@@ -13,6 +13,7 @@ New:
   converted to `JSON`. `infinite` and `nan` floats can be exported using the
   `json5` export format.
 - Added socket API (#2014).
+- Added `liquidsoap.version.at_least`.
 
 Changes:
 - `output.youtube.live` renamed `output.youtube.live.rtmp`, remove `bitrate` and